Update usb-qubes.md
It's unclear to me whether the "Qubes 4.1 only" paragraph applies to (a) the whole "Manual setup for USB keyboards" section or (b) just the "If you wish to use a USB keyboard to enter your LUKS passphrase" part.

My best guess is that the answer is (b). Assuming so, I've proposed adding two additional headings to more clearly set apart the sections about the confirmation prompt and LUKS passphrase options. This would make it clear that the "Qubes 4.1 only" paragraph applies only to the LUKS passphrase option. 

If the answer is actually (a), then of course this proposal should be rejected.